+++ ExperimentalEnemyInteractions/Script.cs
@@ -194,15 +194,16 @@ public class Script : BaseUnityPlugin
         if (BoundingConfig.enableRedBees.Value) Harmony.PatchAll(typeof(BeeAIPatch));
         if (BoundingConfig.enableGiant.Value)Harmony.PatchAll(typeof(ForestGiantPatch));
         if (BoundingConfig.enableSpiderWebs.Value)Harmony.PatchAll(typeof(SandSpiderWebTrapPatch));
+        if (BoundingConfig.enableSpider.Value) Harmony.PatchAll(typeof(SandSpiderAIPatch));
 
         //Compatibilities
-        if (enhancedMonstersPresent && !stableToggle && !enhancedMonstersUseToggle|| enhancedMonstersPresent && enhancedMonstersUseToggle) { Harmony.PatchAll(typeof(EnhancedMonstersCompatibility)); Logger.LogInfo($"Loading compatibility for Enhanced Monsters"); }
-        if (sellBodiesPresent && !stableToggle && !sellBodiesUseToggle || sellBodiesPresent && sellBodiesUseToggle) { SellBodiesFixedCompatibility.AddTracerScriptToPrefabs(); Logger.LogInfo($"Loading compatibility for SellbodiesFixed"); }
+        if (enhancedMonstersPresent && !enhancedMonstersUseToggle|| enhancedMonstersPresent && enhancedMonstersUseToggle) { Harmony.PatchAll(typeof(EnhancedMonstersCompatibility)); Logger.LogInfo($"Loading compatibility for Enhanced Monsters"); }
+        if (sellBodiesPresent && !sellBodiesUseToggle || sellBodiesPresent && sellBodiesUseToggle) { SellBodiesFixedCompatibility.AddTracerScriptToPrefabs(); Logger.LogInfo($"Loading compatibility for SellbodiesFixed"); }
         if (rexuvinationPresent && !rexuvinationUseTogle || rexuvinationPresent && rexuvinationUseTogle) { Harmony.PatchAll(typeof(ReXuvinationPatch)); Logger.LogInfo($"Loading compatibility for Rexuvination"); }
 
-            if (!stableToggle)
+        if (!stableToggle)
         {
-        if (BoundingConfig.enableSpider.Value)Harmony.PatchAll(typeof(SandSpiderAIPatch));
+        //if (BoundingConfig.enableSpider.Value)Harmony.PatchAll(typeof(SandSpiderAIPatch));
 
         if (isExperimental)
         {
@@ -211,10 +212,10 @@ public class Script : BaseUnityPlugin
         }
         else
         {
-                Logger.LogWarning("Limited access. Some patches cannot be enabled in stable branch.");
+        Logger.LogWarning("Limited access. Some patches cannot be enabled in stable branch.");
         }
 
-            Logger.LogInfo("Stable mode off. Loaded all patches.");
+        Logger.LogInfo("Stable mode off. Loaded all patches.");
         }
         else
         {